requirements.txt
GitPython==3.1.30
nornir==3.3.0
nornir-netmiko==0.2.0
nornir-utils==0.2.0

